requirements.txt
# ADR Tool API Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-multipart>=0.0.6
aiofiles>=23.0.0

# AI / LLM Integration
openai>=1.12.0
httpx>=0.26.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
